+++ src/Zhijian/App.axaml.cs
@@ -31,7 +31,9 @@ public partial class App : Application
         if (ApplicationLifetime is IClassicDesktopStyleApplicationLifetime desktop)
         {
             var mainWindow = new MainWindow();
-            mainWindow.DataContext = new MainWindowViewModel(new AvaloniaMindMapFileService(mainWindow));
+            mainWindow.DataContext = new MainWindowViewModel(
+                new AvaloniaMindMapFileService(mainWindow),
+                new AvaloniaApplicationActionService(mainWindow));
             desktop.MainWindow = mainWindow;
         }