+++ src/Zhijian/Services/DisabledApplicationActionService.cs
@@ -1,0 +1,20 @@
+namespace Zhijian.Services;
+
+public sealed class DisabledApplicationActionService : IApplicationActionService
+{
+    public void OpenWebsite()
+    {
+    }
+
+    public void OpenRepository()
+    {
+    }
+
+    public void ShowChangelog()
+    {
+    }
+
+    public void ShowAbout()
+    {
+    }
+}