+++ src/Zhijian/Views/AboutWindow.axaml.cs
@@ -1,0 +1,13 @@
+using Zhijian.ViewModels;
+using AtomWindow = AtomUI.Desktop.Controls.Window;
+
+namespace Zhijian.Views;
+
+public partial class AboutWindow : AtomWindow
+{
+    public AboutWindow()
+    {
+        InitializeComponent();
+        DataContext = new AboutWindowViewModel();
+    }
+}