+++ src/Zhijian/Views/TitleBarLeftAddOn.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace Zhijian.Views;
+
+public partial class TitleBarLeftAddOn : UserControl
+{
+    public TitleBarLeftAddOn()
+    {
+        InitializeComponent();
+    }
+}