+++ src/Zhijian/Views/TitleBarRightAddOn.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace Zhijian.Views;
+
+public partial class TitleBarRightAddOn : UserControl
+{
+    public TitleBarRightAddOn()
+    {
+        InitializeComponent();
+    }
+}